fix: allow decimal values for yearsOfExperience and fix auth for new users
- Change yearsOfExperience from int to float in Pydantic models (backend)
- Parse yearsOfExperience to float in frontend before sending
- Add X-User-Email header fallback for new users without candidate session
- Auto-create candidate record when email fallback is used for new users
- Improve error messages in onboarding form submit catch block

+++ backend/app/api/admin/candidates.py
@@ -33,7 +33,7 @@ class CandidateResponse(BaseModel):
     state: Optional[str] = None
     district: Optional[str] = None
     currentRole: Optional[str] = None
-    yearsOfExperience: Optional[int] = None
+    yearsOfExperience: Optional[float] = None
     farmingBackground: Optional[str] = None
     primaryExpertise: Optional[str] = None
     currentPhase: str
